Change "Service" project to "Services". Add "PointService" class.
